refactor: centralize base URL helper
