+++ src/assets/components/Sliders/ExhibitionPageSlider/ExhibitionPageNewsPopularExhibition.jsx
@@ -14,7 +14,8 @@ import { useNavigate } from 'react-router-dom'
 import { Navigation, Pagination } from 'swiper/modules'
 
 import '@styles/components/Sliders/Base/PopularSlider.scss'
-import { getBaseUrl, getImageUrl } from '../../../../utils/helper'
+import { getBaseUrl } from '../../../../utils/getBaseUrl'
+import { getImageUrl } from '../../../../utils/helper'
 // import LikeAndShare from '../../Blocks/LikeAndShare'
 import TranslatedContent from '../../Blocks/TranslatedContent'
 import Image from 'next/image'
