+++ src/assets/components/Sliders/ExhibitionsPageSlider/ExhibitionsPageTopSlider.jsx
@@ -10,7 +10,8 @@ import 'swiper/css'
 import 'swiper/css/navigation'
 import 'swiper/css/pagination'
 import { Autoplay } from 'swiper/modules'
-import { getBaseUrl, getImageUrl } from '../../../../utils/helper'
+import { getBaseUrl } from '../../../../utils/getBaseUrl'
+import { getImageUrl } from '../../../../utils/helper'
 
 import Image from 'next/image'
