+++ src/assets/components/Sliders/MainPopularArtsSlider/MainPopularArtsSlider.jsx
@@ -13,7 +13,8 @@ import 'swiper/css/pagination'
 import '@styles/components/Sliders/Base/PopularSlider.scss'
 import { useNavigate } from 'react-router-dom'
 import { Navigation, Pagination } from 'swiper/modules'
-import { getBaseUrl, getImageUrl } from '../../../../utils/helper'
+import { getBaseUrl } from '../../../../utils/getBaseUrl'
+import { getImageUrl } from '../../../../utils/helper'
 import ModalWindow from '../../Blocks/ModalWindow'
 import TranslatedContent from '../../Blocks/TranslatedContent'
 import Image from 'next/image'
