+++ src/assets/screens/ExhibitionPage/ExhibitionPage.jsx
@@ -6,7 +6,8 @@ import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useNavigate, useParams } from 'react-router-dom'
 import API from '../../../utils/api.js'
-import { getBaseUrl, getImageUrl } from '../../../utils/helper.js'
+import { getBaseUrl } from '../../../utils/getBaseUrl'
+import { getImageUrl } from '../../../utils/helper.js'
 import LikeAndShare from '../../components/Blocks/LikeAndShare.jsx'
 import TranslatedContent from '../../components/Blocks/TranslatedContent.jsx'
 import ExhibitionPageMasonryGallery from '../../components/Sliders/ExhibitionPageSlider/ExhibitionPageMasonryGallery.jsx'
