+++ src/assets/screens/MuseumPage/MuseumPage.jsx
@@ -6,7 +6,8 @@ import axios from 'axios'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useNavigate, useParams } from 'react-router-dom'
-import { getBaseUrl, getImageUrl } from '../../../utils/helper.js'
+import { getBaseUrl } from '../../../utils/getBaseUrl'
+import { getImageUrl } from '../../../utils/helper.js'
 function MuseumPage() {
 	const { t } = useTranslation()
 	const { id } = useParams()
