+++ src/utils/helper.js
@@ -1,9 +1,4 @@
-export function getBaseUrl() {
-        if (typeof window !== 'undefined') {
-                return window.location.origin
-        }
-        return process.env.NEXT_PUBLIC_API_URL || ''
-}
+import { getBaseUrl } from './getBaseUrl'
 
 /**
  * Gets normalized path to the image.
